# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ads Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt?: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface DashboardSummary {
  cpcAvg: number;
  cpa: number;
  ctr: number;
  roas: number;
  conversions: number;
  totalCost: number;
  cpcChange: number;
  cpaChange: number;
  ctrChange: number;
  roasChange: number;
}

export interface PerformancePoint {
  date: string;
  clicks: number;
  conversions: number;
  cost: number;
}

export interface CampaignConversion {
  name: string;
  value: number;
  percentage: number;
}

export interface Campaign {
  id: number;
  name: string;
  status: string;
  budget: number;
  cpc: number;
  ctr: number;
  roas: number;
  conversions: number;
  createdAt: string;
}

export interface CampaignInput {
  name: string;
  budget: number;
  status?: string;
}

export interface CampaignUpdate {
  name?: string;
  budget?: number;
  status?: string;
}

export interface Keyword {
  id: number;
  keyword: string;
  searchVolume: number;
  competition: string;
  cpc: number;
  location: string;
  period: string;
  /** @nullable */
  analysis?: string | null;
  createdAt: string;
}

export interface KeywordInput {
  keyword: string;
  location?: string;
}

export interface KeywordAnalysisResult {
  id: number;
  analysis: string;
}

export interface TrendPoint {
  month: string;
  volume: number;
}

export interface IntentSlice {
  intent: string;
  percentage: number;
}

export type GetDashboardSummaryParams = {
days?: number;
};

export type GetPerformanceParams = {
days?: number;
};

export type ListKeywordsParams = {
search?: string;
};

export type GetKeywordTrendsParams = {
keyword?: string;
};