# lib/api-zod/src/generated/types/campaign.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ads Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Campaign {
  id: number;
  name: string;
  status: string;
  budget: number;
  cpc: number;
  ctr: number;
  roas: number;
  conversions: number;
  createdAt: string;
}