# lib/api-zod/src/generated/types/campaignConversion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ads Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignConversion {
  name: string;
  value: number;
  percentage: number;
}