# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ads Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  cpcAvg: number;
  cpa: number;
  ctr: number;
  roas: number;
  conversions: number;
  totalCost: number;
  cpcChange: number;
  cpaChange: number;
  ctrChange: number;
  roasChange: number;
}