# lib/api-zod/src/generated/types/keyword.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ads Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Keyword {
  id: number;
  keyword: string;
  searchVolume: number;
  competition: string;
  cpc: number;
  location: string;
  period: string;
  /** @nullable */
  analysis?: string | null;
  createdAt: string;
}