# lib/api-zod/src/generated/types/performancePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ads Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PerformancePoint {
  date: string;
  clicks: number;
  conversions: number;
  cost: number;
}